refactor: Writer/Attendee 도메인의 디렉토리 위치 변경
Plan 도메인에 종속된 서브 도메인으로서 디렉토리의 위치를 변경함.
