+++ src/modules/plans/plans.module.ts
@@ -6,13 +6,12 @@ import { TypeOrmModule } from '@nestjs/typeorm';
 import { Plan } from './entities/plan.entity';
 import { ProjectsModule } from '../projects/projects.module';
 import { AuthModule } from '../auth/auth.module';
-import { Writer } from '../mappings/writers/writers.entity';
 import { UsersModule } from '../users/users.module';
 import { PlanRepository } from './repositories/plan.repository';
 
 @Module({
     imports: [
-        TypeOrmModule.forFeature([Plan, Writer]),
+        TypeOrmModule.forFeature([Plan]),
         forwardRef(() => ProjectsModule),
         AuthModule,
         UsersModule,
