+++ src/modules/users/entities/users.entity.ts
@@ -6,8 +6,8 @@ import { UserProject } from '../../mappings/user-projects/userProjects.entity';
 import { Manager } from '../../mappings/managers/managers.entity';
 import { TaskFile } from '../../mappings/task-files/task-files.entity';
 import { ProjectFile } from '../../mappings/project-files/project-files.entity';
-import { Writer } from '../../mappings/writers/writers.entity';
-import { Attendee } from '../../mappings/attendees/attendees.entity';
+import { Writer } from '../../plans/entities/writers.entity';
+import { Attendee } from '../../plans/entities/attendees.entity';
 import { MasterPortfolio } from '../../master-portfolios/entities/master-portfolios.entity';
 import { Cocomment } from '../../comments/cocomments/entities/cocomments.entity';
 import { PortfolioCorrection } from 'src/modules/portfolio-corrections/entities/portfolio-correction.entity';
